requirements.txt
torch==2.9.1
torchvision==0.24.1
pillow==11.3.0
opencv-python==4.13.0.90
seaborn==0.13.2
numpy==2.2.6
scikit-learn==1.7.2
matplotlib==3.10.6